requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0              # process manager (Gunicorn + UvicornWorker)
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12

# ── Anthropic ─────────────────────────────────────────────────────────────────
anthropic==0.40.0

# ── Google ────────────────────────────────────────────────────────────────────
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# ── Twilio ────────────────────────────────────────────────────────────────────
twilio==9.3.4

# ── Slack ─────────────────────────────────────────────────────────────────────
slack-sdk==3.33.0

# ── Database ──────────────────────────────────────────────────────────────────
asyncpg==0.29.0

# ── Redis + Celery ────────────────────────────────────────────────────────────
redis==5.1.1
celery[redis]==5.4.0
celery-redbeat==2.2.0         # Redis-backed beat scheduler (survives restarts)

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0

# ── Encryption ────────────────────────────────────────────────────────────────
cryptography==43.0.1

# ── Browser automation ────────────────────────────────────────────────────────
playwright==1.48.0

# ── Embeddings (semantic memory) ──────────────────────────────────────────────
openai==1.54.0

# ── Monitoring & Observability ────────────────────────────────────────────────
sentry-sdk[fastapi]==2.17.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-exporter-otlp==1.27.0

# ── CLI ───────────────────────────────────────────────────────────────────────
httpx==0.27.2
rich==13.9.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
factory-boy==3.3.1

# ── Load testing ──────────────────────────────────────────────────────────────
locust==2.32.2
